// src/utils/enums/common.ts
/* eslint-disable no-unused-vars */
/**
 * Enum to define user roles
 * @enum {String}
 * @readonly
 */
export enum Roles {
  ADMIN = 'Administrator',
  CASHIER = 'Cashier',
  CUSTOMER = 'Customer'
}

/**
 * Enum to define the periods to filter dashboard
 * @enum {String}
 * @readonly
 */
export enum Period {
  daily = 'Daily',
  weekly = 'Weekly',
  monthly = 'Monthly'
}